kernel: Remove BORE note
I believe that this clutters the page and is somehow redundant information. We already have hyperlinks that leads to
BORE's GitHub repository to explain what exactly BORE is, and I find that better than a oneliner note that we had.
If we want, we should explain better what BORE is and how different it works compared to EEVDF.

Signed-off-by: Eric Naim <dnaim@cachyos.org>