various: Remove ambiguous personal pronouns
The usage of "we", "us", "you" is ambiguous to the users reading. Use CachyOS directly where necessary and more generic
subjects.

There's more of this than I thought that this could potentially block this PR so I'm going to do the rest at a later time.
Some pages (install) are also halfheartedly modified for now because I want to rewrite some of the sections eventually
and this can cause double work (unacceptable for a lazy bum like me :D).

Signed-off-by: Eric Naim <dnaim@cachyos.org>